Move JobApplication details content and info header to admin's components
