+++ src/pages/Admin/components/JobApplication/DetailContent/component.tsx
