+++ src/pages/Admin/components/JobApplication/DetailContent/index.ts
