+++ src/pages/Admin/components/JobApplication/DetailContent/interfaces.ts
