+++ src/pages/Admin/components/JobApplication/DetailContent/styles.module.scss
