+++ src/pages/Admin/components/JobApplication/DetailInfo/MainTitle/index.ts
