+++ src/pages/Admin/components/JobApplication/DetailInfo/component.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from "react";
 import { ApprovalStatus } from "$interfaces/ApprovalStatus";
 import { IJobApplication } from "$interfaces/JobApplication";
-import { DetailInfo } from "../../../../../components/DetailInfo";
-import { UserDetails } from "../../../../../components/Applicant/ApplicantDetailInfo/UserDetails";
+import { DetailInfo } from "../../DetailInfo";
+import { UserDetails } from "../../Applicant/ApplicantDetailInfo/UserDetails";
 import { MainTitle } from "./MainTitle";
 
 export const JobApplicationDetailInfo: FunctionComponent<IComponentProps> = ({
