+++ src/pages/Admin/components/JobApplication/DetailInfo/index.ts
